+++ src/Sql/Sql/Database Backup/Model/AzureSqlDatabaseLongTermRetentionBackupModel.cs
@@ -75,11 +75,6 @@ namespace Microsoft.Azure.Commands.Sql.Backup.Model
         /// <summary>
         /// Gets or sets the current backup storage redundancy for the database
         /// </summary>
-        public string CurrentBackupStorageRedundancy { get; set; }
-
-        /// <summary>
-        /// Gets or sets the requested backup storage redundancy for the database
-        /// </summary>
-        public string RequestedBackupStorageRedundancy { get; set; }
+        public string BackupStorageRedundancy { get; set; }
     }
 }